逻辑调整

+++ src/main/java/com/example/sbimgutil/task/BaseTask.java
@@ -4,13 +4,11 @@ package com.example.sbimgutil.task;
 import cn.hutool.core.date.LocalDateTimeUtil;
 import com.example.sbimgutil.context.TaskExcutor;
 import com.example.sbimgutil.schedule.ITask;
+import com.example.sbimgutil.schedule.TaskStateEnum;
 import com.example.sbimgutil.utils.ConsoleProgressBar;
 import com.example.sbimgutil.utils.Const;
-import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.imageio.ImageIO;
 import javax.imageio.ImageWriter;
@@ -32,30 +30,23 @@ public abstract class BaseTask implements ITask {
         ImageWriter writer = writers.next();
     }
 
-    public static final FileFilter SUPPORTED_FILE_FILTER = file -> {
-        if (file.isDirectory())
-            return true;
-        String lowerCasedName = file.getName().toLowerCase();
-        return Const.SUPORTTED_FORMATS.stream().anyMatch(lowerCasedName::endsWith);
-    };
-
-
     private LocalDateTime startDate;
-
     protected String taskName;
     protected TaskStateEnum state = TaskStateEnum.NEW;
-
     protected File outFile;
 
     @Override
     public void before() throws IOException {
-        outFile = new File(outFile.getParentFile(), outFile.getName() + ".tmp");
-        if (outFile.exists()) {
-            Files.delete(outFile.toPath());
+        if (outFile != null) {
+            outFile = new File(outFile.getParentFile(), outFile.getName() + ".tmp");
+            if (outFile.exists()) {
+                Files.delete(outFile.toPath());
+            }
+            if (!outFile.getParentFile().exists()) {
+                FileUtils.forceMkdirParent(outFile);
+            }
         }
-        if (!outFile.getParentFile().exists()) {
-            FileUtils.forceMkdirParent(outFile);
-        }
+
         state = TaskStateEnum.RUNNING;
         startDate = LocalDateTime.now();
     }
@@ -63,11 +54,12 @@ public abstract class BaseTask implements ITask {
 
     @Override
     public void after() {
-        outFile.renameTo(new File(outFile.getParentFile(), outFile.getName().substring(0, outFile.getName().lastIndexOf("."))));
+        if (outFile != null && outFile.exists())
+            outFile.renameTo(new File(outFile.getParentFile(), outFile.getName().substring(0, outFile.getName().lastIndexOf("."))));
 
         long between = LocalDateTimeUtil.between(startDate, LocalDateTime.now(), ChronoUnit.SECONDS);
-        log.debug("任务完成:{},执行时间：{}s", taskName, between);
-        state = TaskStateEnum.TERMINATED;
+        log.debug("任务完成:{},执行耗时：{}s", taskName, between);
+        state = TaskStateEnum.FINISHED;
 
         ConsoleProgressBar progressBar = TaskExcutor.getGlobalConsoleProgressBar();
         if (progressBar != null) {
